+++ app/src/main/java/at/aau/gloryweapons/siegeanddestroy3d/network/interfaces/NetworkCommunicatorServer.java
@@ -2,6 +2,8 @@ package at.aau.gloryweapons.siegeanddestroy3d.network.interfaces;
 
 import android.app.Activity;
 
+import java.util.List;
+
 public interface NetworkCommunicatorServer {
     /**
      * Initialization of the server
@@ -9,13 +11,14 @@ public interface NetworkCommunicatorServer {
      * @param activity     current activity
      * @param userCallBack a callback for the representation of the clients
      */
-    public void initServerGameHandler(final Activity activity, UserCallBack userCallBack);
+    public void initServerGameHandler(final Activity activity, CallbackObject<List<String>> userCallBack);
 
     /**
      * Stop network and disable wifi direct
      */
     public void resetNetwork();
 
+    // TODO move to GlobalGameConfig
     /**
      * @param shotCount
      */
