--- app/src/main/java/at/aau/gloryweapons/siegeanddestroy3d/network/interfaces/UserCallBack.java
@@ -1,8 +1,0 @@
-package at.aau.gloryweapons.siegeanddestroy3d.network.interfaces;
-
-import java.util.List;
-
-// TODO remove useless interface + references
-public interface UserCallBack extends CallbackObject<List<String>> {
-
-}
