+++ app/src/main/java/at/aau/gloryweapons/siegeanddestroy3d/network/kryonet/ClientGameHandlerKryoNet.java
@@ -3,17 +3,6 @@ package at.aau.gloryweapons.siegeanddestroy3d.network.kryonet;
 import android.app.Activity;
 import android.util.Log;
 
-import com.koushikdutta.async.AsyncServer;
-import com.koushikdutta.async.AsyncSocket;
-import com.koushikdutta.async.ByteBufferList;
-import com.koushikdutta.async.DataEmitter;
-import com.koushikdutta.async.Util;
-import com.koushikdutta.async.callback.CompletedCallback;
-import com.koushikdutta.async.callback.ConnectCallback;
-import com.koushikdutta.async.callback.DataCallback;
-
-import java.net.InetSocketAddress;
-
 import com.esotericsoftware.kryonet.Client;
 import com.esotericsoftware.kryonet.Connection;
 import com.esotericsoftware.kryonet.Listener;
@@ -35,9 +24,9 @@ import at.aau.gloryweapons.siegeanddestroy3d.network.dto.UserNameRequestDTO;
 import at.aau.gloryweapons.siegeanddestroy3d.network.dto.UserNameResponseDTO;
 import at.aau.gloryweapons.siegeanddestroy3d.network.dto.WrapperHelper;
 import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.CallbackObject;
-import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.NetworkCommunicator;
+import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.NetworkCommunicatorClient;
 
-public class ClientGameHandlerKryoNet implements NetworkCommunicator {
+public class ClientGameHandlerKryoNet implements NetworkCommunicatorClient {
 
     private Client kryoClient;
     private KryonetHelper kryoHelper;
