+++ app/src/main/java/at/aau/gloryweapons/siegeanddestroy3d/network/kryonet/ServerGameHandlerKryoNet.java
@@ -1,25 +1,6 @@
 package at.aau.gloryweapons.siegeanddestroy3d.network.kryonet;
 
 import android.app.Activity;
-import android.net.wifi.WifiManager;
-import android.text.format.Formatter;
-import android.util.Log;
-
-import com.koushikdutta.async.AsyncServer;
-import com.koushikdutta.async.AsyncServerSocket;
-import com.koushikdutta.async.AsyncSocket;
-import com.koushikdutta.async.ByteBufferList;
-import com.koushikdutta.async.DataEmitter;
-import com.koushikdutta.async.DataSink;
-import com.koushikdutta.async.Util;
-import com.koushikdutta.async.callback.CompletedCallback;
-import com.koushikdutta.async.callback.DataCallback;
-import com.koushikdutta.async.callback.ListenCallback;
-import com.koushikdutta.async.callback.WritableCallback;
-
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.nio.ByteBuffer;
 import android.util.Log;
 
 import com.esotericsoftware.kryonet.Connection;
@@ -44,14 +25,11 @@ import at.aau.gloryweapons.siegeanddestroy3d.network.dto.UserNameRequestDTO;
 import at.aau.gloryweapons.siegeanddestroy3d.network.dto.UserNameResponseDTO;
 import at.aau.gloryweapons.siegeanddestroy3d.network.dto.WrapperHelper;
 import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.CallbackObject;
-import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.NetworkCommunicator;
+import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.NetworkCommunicatorClient;
 import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.NetworkCommunicatorServer;
-import at.aau.gloryweapons.siegeanddestroy3d.network.interfaces.UserCallBack;
 import at.aau.gloryweapons.siegeanddestroy3d.server.ServerController;
 
-import static android.content.Context.WIFI_SERVICE;
-
-public class ServerGameHandlerKryoNet implements NetworkCommunicatorServer, NetworkCommunicator {
+public class ServerGameHandlerKryoNet implements NetworkCommunicatorServer, NetworkCommunicatorClient {
     private Server kryoServer;
     private KryonetHelper kryoHelper;
 
@@ -62,7 +40,7 @@ public class ServerGameHandlerKryoNet implements NetworkCommunicatorServer, Netw
     private HashMap<Integer, ClientData> clientDataMap;
 
     //client name list
-    private UserCallBack userCallBack;
+    private CallbackObject<List<String>> userCallBack;
 
     CallbackObject<User> turnInfoUpdateCallback;
 
@@ -84,7 +62,7 @@ public class ServerGameHandlerKryoNet implements NetworkCommunicatorServer, Netw
     }
 
     @Override
-    public void initServerGameHandler(Activity activity, UserCallBack userCallBack) {
+    public void initServerGameHandler(Activity activity, CallbackObject<List<String>> userCallBack) {
         GlobalGameSettings.getCurrent().setServer(true);
 
         wrapperHelper = WrapperHelper.getInstance();
